fix: favourite should update cache
Toggling a series as favourite will now properly reflect this in the cache.
Previously once the anime details were closed and reopened the previous value would be shown instead of the new favourite value.
Favourite icon is now colored when a series is favourited, this should make it easier to distinguish a favourited series.