feat: dynamic extension loading
Heavily reworked extensions, they now load dynamically so search results will appear as each extension completes its query significantly decreasing wait times.

Extension queries are now cached for a little over 1 minute so closing the torrent search then re-opening it wont result in another long wait for results.

Torrent Search has been reworked to be clean and minimal with descriptive details. The auto-selected torrent is clearly visible and autoplay indicator is vibrant. If the previously selected torrent is the same as the auto-selected torrent they will be merged for clarity.

Trusted sources are now visible under the extensions section in the settings when clicking to add a source.