feat: video file manager
Adds a button in the player to open the File Manager where you can view all playable files for the current torrent, manually adjust the resolved series if its incorrect, as well as adjust the episode number of each file. This view can also be opened in the player via the `H` keybind.
Torrent hashes and file hashes are now mapped to the resolved media, allowing for easy tracking of downloads.
Adds pre-download button to episode cards, and notification cards.
Now automatically plays the torrent if the matching series is found in the torrent files instead of prompting the torrent modal. This also overrides the RSS feed magnets, unless that magnet is already mapped to a file.