chore: modified prefer dubs
Prefer dubs will now only function so long as your current progress is not greater than the last released dub episode. If your progress exceeds this it will be assumed that you are watching the series subbed.

This works well when you want to watch some series dubbed and some series (that have a dub) subbed.