chore: simplify and cleanup webtorrent
Improves performance and discards the reliance on skipVerify to better handle torrents that could be corrupted.
Implements "missing" UI feedback when pieces for a complete torrent exist but are missing others.
Properly locks rescan until torrents are loaded on startup and during a rescan.
Pre-download and rescan are now completely disabled when persist files is disabled and the seeding limit is 1.