chore: improve notifications performance
Greatly improves the performance when sending MANY notifications, typically when starting up the app after not being used for several days.
Reduces the number of system notifications sent to the 20 most recent (in each batch) which silences noisy notifications. The in-app "local" notifications are not affected by this.
New notifications are processed in batches improving performance.
Better deduplication when handling batch notifications, prevents the system from being spammed with RSS and Schedule notifications when RSS takes priority.