chore: improve player to external player
Heavily improves the external player functionality in relation to the built-in player. Functionality of the external player is triggered similar to how the built-in player functions.
External player now only appears when you click to play the video (or if auto play is enabled in settings) instead of force opening the external player as soon as the video loads.
Buttons that do not relate to the external player will now be hidden from the built-in player.