chore: improve zero episode management
Fixes edge cases with zero episodes affecting the file manager.
Zero episodes will now be displayed and allow the episode number to be changed.
A series with more than 1 episode but the file is missing its episode number will automatically show the episode number as zero so it can be changed.
Changing a series that has a zero episode will change the target file if its a zero episode without modifying the others.